% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{date_to_mongo}
\alias{date_to_mongo}
\title{translate a date into mongo numberLong query}
\usage{
date_to_mongo(datestring, format = "\%Y-\%m-\%d")
}
\description{
translate a date into mongo numberLong query
}
\examples{
date_to_mongo('2020-03-18')
# { "$date" : { "$numberLong" : "1584514800000" } }
}
