% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{find_odbc_by_server}
\alias{find_odbc_by_server}
\title{finds Windows ODBC credentials assuming the REGISTRY_FILE exists
breaks if file is missing (to be caught upstream)}
\usage{
find_odbc_by_server(server, REGISTRY_FILE = "Software\\\\ODBC\\\\ODBC.INI")
}
\description{
finds Windows ODBC credentials assuming the REGISTRY_FILE exists
breaks if file is missing (to be caught upstream)
}
