% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{retrieves or requests and stores credentials as necessary for ODBC connection}
\usage{
get_credentials(
  server = NULL,
  reset = F,
  urlencode = F,
  forcekeyring = F,
  domain = T
)
}
\arguments{
\item{server}{server name (or any other identifier string that identifies username--password pair}

\item{reset}{whether to reset credentials should they be already recorded in the registry}

\item{urlencode}{URL-encode the password}

\item{forcekeyring}{force keyring credential storage on Windows (on Mac and Unix keyring is used by default)}

\item{domain}{require username to have a domain prefix ("domain\\username")}
}
\value{
list(uid="myusername", pwd="password") # on Mac / Unix and list(uid=NULL, pwd=NULL) # on Windows
}
\description{
On Mac and Unix, credentials are stored in keyring
(can be forced on Windows with forcekeyring=F)
}
\details{
On Windows, ODBC Data Source Administrator is called
(no credentials are explicitly retrieved -- list with NULL entries is returned
}
