% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials_direct}
\alias{get_credentials_direct}
\title{request credentials or retrieve them from system's keyring
if reset=T -- clears the keyring}
\usage{
get_credentials_direct(dbname, uid = NULL, reset = F, domain = T)
}
\description{
request credentials or retrieve them from system's keyring
if reset=T -- clears the keyring
}
