% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials_windows}
\alias{get_credentials_windows}
\title{fetch / register credentials in the ODBC Data Source Administrator}
\usage{
get_credentials_windows(
  server,
  reset = F,
  REGISTRY_FILE = "Software\\\\ODBC\\\\ODBC.INI"
)
}
\description{
fetch / register credentials in the ODBC Data Source Administrator
}
