% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_mssql_driver}
\alias{get_mssql_driver}
\title{retrieves default driver}
\usage{
get_mssql_driver(dbconfig = list())
}
\arguments{
\item{dbconfig}{configuration list (optional)}
}
\value{
dbconfig    configuration list with driver name appended: list(driver="drivername", ...)
}
\description{
retrieves default driver
(Mac / Unix: FreeTDS, Windows: SQL Server)
}
\examples{
dbconfig <- read_yaml("database_info.yaml")
dbconfig <- get_mssql_driver(dbconfig)
credentials <- get_credentials(dbconfig$server)

connection <- DBI::dbConnect(
    odbc(),
    driver=dbconfig$driver,
    server=dbconfig$server,
    port=dbconfig$port,
    uid=credentials$uid,
    pwd=credentials$pwd,
    database='mydatabase'
    )
}
