% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{mj}
\alias{mj}
\title{query clean-up and substitution for mongoDB}
\usage{
mj(txt, replace = T, .open = "{{", .close = "}}")
}
\description{
cleans up mongoDB JSON to standard JSON and
perform \code{glue} replacements with double braces
}
\examples{
# enclose variables that need to be substituted with double curvy braces, e.g.
term <- "tree"
mj('{$search: "{{term}}"')
# {"$search": "tree"}
}
